pyproject.toml
[tool.poetry]
name = "prts"
version = "0.1.0"
description = "prts"
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.7"
nonebot2 = "^2.0.0.a1"
nb-cli = "^0.1.0"
docker-compose = "^1.28.0"
bs4 = "^0.0.1"
httpx = "^0.16.1"
nonebot-plugin-apscheduler = "^0.1.2"

[tool.poetry.dev-dependencies]
nonebot-test = "^0.1.0"
pylint = "^2.6.0"
jedi = "^0.18.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
